+++ apps/www/src/modules/common/components/marquee.tsx
@@ -1,0 +1,63 @@
+"use client";
+
+import {
+  motion,
+  useAnimationFrame,
+  useMotionValue,
+  useScroll,
+  useSpring,
+  useTransform,
+  useVelocity,
+  wrap,
+} from "framer-motion";
+import { useRef } from "react";
+
+interface MarqueeProps {
+  children: string;
+  baseVelocity?: number;
+}
+
+const Marquee = ({ children, baseVelocity = 5 }: MarqueeProps) => {
+  const baseX = useMotionValue(0);
+  const { scrollY } = useScroll();
+  const scrollVelocity = useVelocity(scrollY);
+  const smoothVelocity = useSpring(scrollVelocity, {
+    damping: 50,
+    stiffness: 400,
+  });
+  const velocityFactor = useTransform(smoothVelocity, [0, 1000], [0, 5], {
+    clamp: false,
+  });
+
+  const x = useTransform(baseX, (v) => `${wrap(-20, -45, v)}%`);
+
+  const directionFactor = useRef<number>(1);
+
+  useAnimationFrame((t, delta) => {
+    let moveBy = directionFactor.current * baseVelocity * (delta / 1000);
+
+    if (velocityFactor.get() < 0) {
+      directionFactor.current = -1;
+    } else if (velocityFactor.get() > 0) {
+      directionFactor.current = 1;
+    }
+
+    moveBy += directionFactor.current * moveBy * velocityFactor.get();
+
+    baseX.set(baseX.get() + moveBy);
+  });
+
+  return (
+    <>
+      <motion.div style={{ x }} className="flex whitespace-nowrap gap-14">
+        {Array.from({ length: 4 }, () => children).map((item) => (
+          <span className="text-[140px] leading-[140px] font-bold text-dark uppercase">
+            {item}
+          </span>
+        ))}
+      </motion.div>
+    </>
+  );
+};
+
+export { Marquee };