+++ apps/www/src/modules/common/widgets/footer.tsx
@@ -4,6 +4,8 @@ import Link from "next/link";
 import { Button } from "../components/button";
 import useFooter from "../hooks/useFooter";
 import { Separator } from "../components/separator";
+import Experience from "../components/experience";
+import Clients from "@/modules/home/widgets/clients";
 
 const socials = [
   {
@@ -27,6 +29,8 @@ const Footer = () => {
   return (
     <>
       <Separator className="my-32" container />
+      <Clients container className="mt-28" />
+      <Experience />
       <BaseFooter />
     </>
   );